split cell types code
